Update get_rate function for international letter rates
Letter rates in the database can be `first`, `second` or
`international` - we don't have separate rates for letters going to
Europe or the rest of the world. Postage on notifications can have the
value of `europe` or `rest-of-world`, so the `get_rate` function maps
both these postage types to `international`.